api/requirements.txt
anyio==3.6.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
easyocr==1.6.2
fastapi==0.85.0
h11==0.14.0
httptools==0.5.0
idna==3.4
imageio==2.22.1
networkx==2.8.7
ninja==1.10.2.4
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pyclipper==1.3.0.post3
pydantic==1.10.2
pyparsing==3.0.9
python-bidi==0.4.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.2
Shapely==1.8.4
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tifffile==2022.8.12
torch==1.12.1
torchvision==0.13.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
opencv-python-headless
